// src/actions/types.ts
export enum ActionsEnum {
    INIT_APP = 'INIT_APP',

    SEND_MESSAGE = 'SEND_MESSAGE',
    SEND_SUCCESS = 'SEND_SUCCESS',
    SEND_FAILED = 'SEND_FAILED',
    SENDING = 'SENDING',

    SET_CONTENT = 'SET_CONTENT',
    CONTENT_FETCH_FAILED = 'CONTENT_FETCH_FAILED',

    LOAD_ARTICLES = 'LOAD_ARTICLES',
    FAILED_LOADING_ARTICLES = 'FAILED_LOADING_ARTICLES',
    SET_ARTICLES = 'SET_ARTICLES',

    LOAD_ARTICLE = 'LOAD_ARTICLE',
    FAILED_LOADING_ARTICLE = 'FAILED_LOADING_ARTICLE',
    SET_ARTICLE = 'SET_ARTICLE',

    SET_LANGUAGE = 'SET_LANGUAGE',
    TOGGLE_LANGUAGE = 'TOGGLE_LANGUAGE',
}

export default ActionsEnum;